Added Discord Bot Python Script
This script will be run on any server.  In my case, I run it on a Linux server as a systemd service.  This way it runs all the time and at server boot time.

This script constantly talks to the Discord channel ID, and keeps your bot active to watch for uploads of N.I.N.A. AutoFocus JSON files.